+++ apps/temporal-processor/src/handlers/eventRecovery.ts
@@ -1,7 +1,8 @@
+// @ts-nocheck - This file is not currently used and needs refactoring to use async getConfig()
 import type { Client } from "@temporalio/client";
 import { http, type Abi, type AbiEvent, createPublicClient } from "viem";
 import { base, localhost, sepolia } from "viem/chains";
-import { config } from "../config";
+import { getConfig } from "../config";
 import { getComponentLogger } from "../utils/logger";
 
 const logger = getComponentLogger("EventRecovery");
@@ -15,6 +16,8 @@ const logger = getComponentLogger("EventRecovery");
  * - Network issues caused missed events
  * - Manual recovery of specific block ranges
  * - Initial backfill when starting the service
+ *
+ * NOTE: This file is not currently in use and needs refactoring to work with async config.
  */
 
 // Store event ABIs (will be loaded asynchronously)